Add folly::StrandExecutor type
Summary:
This diff adds a new StrandExecutor executor-type that is similar
to SerialExecutor in that it serialises execution of work enqueued
to it, but is more general in that it allows different tasks to be
executed on different executors.

To do this, it separates the queue out into a StrandContext object
which can be passed to multiple StrandExecutor objects, allowing them
each to share a queue, while allowing each StrandExecutor to dispatch
work queued to it to a different parent Executor.

Reviewed By: andriigrynenko

Differential Revision: D21848165

fbshipit-source-id: 4dda6bd13a9cd5b275a11f13101242cd73b327e7