clang coro handling
Summary:
On windows `<experimental/coroutine>` comes from the Microsoft SDK and uses the MSVC intrinsics. If you are compiling with clang this header exists but nothing from it will work because clang does not implement the MSVC intrinsics.

Further, the clang and MSVC intrinsics are ABI incompatible with each other to the extent that there is no way I can find to implement the MSVC intrinsics behavior as inline functions in a clang TU.

MS have indicated publicly that they will work with the LLVM project to make sure that clang and MSVC will be compatible for `<coroutine>`, but until then we have to say that folly does not have coroutines when clang is used on Windows.

(Note: this ignores all push blocking failures!)

Reviewed By: yfeldblum

Differential Revision: D22301662

fbshipit-source-id: 96b1c3909a3916fe300073af34429eeb5c08d1fd