Attempt to fix a race condition around fiber stats
Summary: Mcrouter has a logger thread which queries the fiber manager for 3 stats; these stats are non-atomic and therefore this causes a race. This diff applies a simple fix to make these atomic. It does not feel like this is the best fix though, so let me know if anyone has ideas for a better fix. It feels that stats should be provided some other way (rather than a bunch of atomics)...

Reviewed By: yfeldblum

Differential Revision: D22878457

fbshipit-source-id: f17544b38836c454162b83cfa570c90c5e7e6b9a