Avoid std::string with reentrant_allocator in ElfCache
Summary:
The implementation of `std::basic_string` with the legacy ABI and with `_GLIBCXX_FULLY_DYNAMIC_STRING`, as appears by default on CentOS 7, incorrectly requires the provided allocator type to have a default constructor. The `Allocator` concept does not require this and many allocator types with runtime state shared between copies will not provide this.

Switch to `folly::fbstring` in `ElfCache` to sidestep this problem.

Fixes: https://github.com/facebook/folly/issues/1344.

Reviewed By: yfeldblum

Differential Revision: D22970418

fbshipit-source-id: 1e188d668991d25f53da5702b31f4a7a7ec7d7e8